requirements.txt
tensorflow==2.8.0
streamlit==0.62.0
numpy==1.16.5
pillow==6.2.0